move prop_all_common_prefix to Test.Dynamic.Util
