+++ ouroboros-consensus/test-consensus/Test/Dynamic/Praos.hs
@@ -14,19 +14,13 @@
 
 module Test.Dynamic.Praos (
     tests
-  , prop_all_common_prefix
   ) where
 
 import qualified Data.Map.Strict as Map
-import           Data.Word (Word64)
 import           Test.QuickCheck
 
 import           Test.Tasty
 import           Test.Tasty.QuickCheck
-
-import           Ouroboros.Network.Block
-import           Ouroboros.Network.MockChain.Chain
-import qualified Ouroboros.Network.MockChain.Chain as Chain
 
 import           Ouroboros.Consensus.BlockchainTime
 import           Ouroboros.Consensus.Demo
@@ -38,7 +32,6 @@ import           Ouroboros.Consensus.Util.Random
 import           Test.Dynamic.General
 import           Test.Dynamic.Util
 
-import           Test.Util.MockChain (dropLastBlocks, lastSlot)
 import           Test.Util.Orphans.Arbitrary ()
 import           Test.Util.Range
 
@@ -101,42 +94,3 @@ prop_simple_praos_convergence
     schedule = leaderScheduleFromTrace numSlots final
     longest  = longestCrowdedRun schedule
     crowded  = crowdedRunLength longest
-
-prop_all_common_prefix :: (HasHeader b, Condense b, Eq b)
-                       => Word64 -> [Chain b] -> Property
-prop_all_common_prefix _ []     = property True
-prop_all_common_prefix l (c:cs) = conjoin [prop_common_prefix l c d | d <- cs]
-
-prop_common_prefix :: forall b. (HasHeader b, Condense b, Eq b)
-                   => Word64 -> Chain b -> Chain b -> Property
-prop_common_prefix l x y = go x y .&&. go y x
-  where
-    go c d =
-        let (l', c') = findPrefix c d
-            e        = "after dropping "
-                 <> show l'
-                 <> " blocks from "
-                 <> showChain c
-                 <> ",\n\nthe resulting "
-                 <> showChain c'
-                 <> "\n\nis a prefix of "
-                 <> showChain d
-                 <> ",\n\nbut only "
-                 <> show l
-                 <> " block(s) should have been necessary"
-        in  counterexample e $ l' <= l
-
-    findPrefix c' d
-        | c' `isPrefixOf` d = (0, c')
-        | otherwise         = let (l', c'') = findPrefix (dropLastBlocks 1 c') d
-                              in  (l' + 1, c'')
-
-    showChain :: Chain b -> String
-    showChain c = condense c
-                  <> "\n(length "
-                  <> show (Chain.length c)
-                  <> case lastSlot c of
-                        Nothing -> ")"
-                        Just s  ->    ", last slot "
-                                   <> show (unSlotNo s)
-                                   <> ")"
