+++ ouroboros-consensus/test-consensus/Test/Dynamic/Util.hs
@@ -6,6 +6,7 @@
 
 module Test.Dynamic.Util (
     allEqual
+  , prop_all_common_prefix
   , shortestLength
   , tracesToDot
   , leaderScheduleFromTrace
@@ -46,6 +47,10 @@ import           Ouroboros.Consensus.Util.Orphans ()
 import           Test.Util.HasCreator
 import qualified Test.Util.MockChain as Chain
 
+{-------------------------------------------------------------------------------
+  Chain properties
+-------------------------------------------------------------------------------}
+
 allEqual :: forall b. (Condense b, Eq b, HasHeader b) => [Chain b] -> Property
 allEqual []             = property True
 allEqual [_]            = property True
@@ -81,6 +86,45 @@ allEqual (x : xs@(_:_)) =
 
 shortestLength :: Map NodeId (Chain b) -> Natural
 shortestLength = fromIntegral . minimum . map Chain.length . Map.elems
+
+prop_all_common_prefix :: (HasHeader b, Condense b, Eq b)
+                       => Word64 -> [Chain b] -> Property
+prop_all_common_prefix _ []     = property True
+prop_all_common_prefix l (c:cs) = conjoin [prop_common_prefix l c d | d <- cs]
+
+prop_common_prefix :: forall b. (HasHeader b, Condense b, Eq b)
+                   => Word64 -> Chain b -> Chain b -> Property
+prop_common_prefix l x y = go x y .&&. go y x
+  where
+    go c d =
+        let (l', c') = findPrefix c d
+            e        = "after dropping "
+                 <> show l'
+                 <> " blocks from "
+                 <> showChain c
+                 <> ",\n\nthe resulting "
+                 <> showChain c'
+                 <> "\n\nis a prefix of "
+                 <> showChain d
+                 <> ",\n\nbut only "
+                 <> show l
+                 <> " block(s) should have been necessary"
+        in  counterexample e $ l' <= l
+
+    findPrefix c' d
+        | c' `Chain.isPrefixOf` d = (0, c')
+        | otherwise         = let (l', c'') = findPrefix (Chain.dropLastBlocks 1 c') d
+                              in  (l' + 1, c'')
+
+    showChain :: Chain b -> String
+    showChain c = condense c
+                  <> "\n(length "
+                  <> show (Chain.length c)
+                  <> case Chain.lastSlot c of
+                        Nothing -> ")"
+                        Just s  ->    ", last slot "
+                                   <> show (unSlotNo s)
+                                   <> ")"
 
 {-------------------------------------------------------------------------------
   Generation of a dot-file to represent the trace as a graph
